+++ src/Formatter/DefaultFormatter.php
@@ -81,6 +81,10 @@ class DefaultFormatter extends AbstractFormatter implements FormatterInterface
 
     public function getLabelProperties(ReflectionClass $reflection): string
     {
+        if (!$this->options['show-properties']) {
+            return '';
+        }
+
         $properties = $reflection->getProperties();
 
         if (count($properties) === 0) {
@@ -136,6 +140,10 @@ class DefaultFormatter extends AbstractFormatter implements FormatterInterface
 
     public function getLabelFunctions(array $functions, string $class = null): string
     {
+        if ($class && !$this->options['show-methods']) {
+            return '';
+        }
+
         $operations = [];
 
         foreach ($functions as $method) {